AP_Camera: changed "int" to "int16_t" and "long" to "int32_t" here and there.

+++ libraries/AP_Camera/AP_Camera.h
@@ -36,15 +36,15 @@ public:
 	void configure_msg(mavlink_message_t* msg);
 	void control_msg(mavlink_message_t* msg);
 
-	int picture_time;					///< waypoint trigger variable
-	long wp_distance_min;				///< take picture if distance to WP is smaller than this
+	int16_t picture_time;					///< waypoint trigger variable
+	int32_t wp_distance_min;				///< take picture if distance to WP is smaller than this
 
     static const struct AP_Param::GroupInfo var_info[];
 
 private:
 	uint8_t keep_cam_trigg_active_cycles; ///< event loop cycles to keep trigger active
-	int thr_pic;						///< timer variable for throttle_pic
-	int camtrig;						///< PK6 chosen as it not near anything so safer for soldering
+	int16_t thr_pic;						///< timer variable for throttle_pic
+	int16_t camtrig;						///< PK6 chosen as it not near anything so safer for soldering
 
 	AP_Int8		trigger_type;	///< 0=Servo, 1=relay, 2=throttle_off time, 3=throttle_off waypoint, 4=transistor
